// src/types/actions/transactionActionTypes.ts
export enum TRANSACTIONS_ACTION_TYPES {
  FETCH_START = 'FETCH_START',
  FETCH_SUCCESS = 'FETCH_SUCCESS',
  FETCH_ERROR = 'FETCH_ERROR',
  EDIT_START = 'EDIT_START',
  CREATE_SUCCESS = 'CREATE_SUCCESS',
  EDIT_ERROR = 'EDIT_ERROR',
  REMOVE_SUCCESS = 'REMOVE_SUCCESS',
  EDIT_SUCCESS = 'EDIT_SUCCESS',
}